+++ src/cell_draw.cpp
@@ -1183,8 +1183,8 @@ void cell_draw()
 
     if (gdata_torch == 1)
     {
-        if (mdata_map_type >= mdata_t::MapType::dungeon
-            && mdata_map_type <= mdata_t::MapType::dungeon_castle)
+        if (mdata_map_type >= static_cast<int>(mdata_t::MapType::dungeon)
+            && mdata_map_type <= static_cast<int>(mdata_t::MapType::dungeon_castle))
         {
             light_ -= 50;
         }
