+++ src/mapgen.cpp
@@ -1378,8 +1378,8 @@ void map_randsite(int prm_971, int prm_972)
     {
         return;
     }
-    if (mdata_map_type >= mdata_t::MapType::dungeon
-        && mdata_map_type <= mdata_t::MapType::dungeon_castle)
+    if (mdata_map_type >= static_cast<int>(mdata_t::MapType::dungeon)
+        && mdata_map_type <= static_cast<int>(mdata_t::MapType::dungeon_castle))
     {
         if (mdata_map_next_regenerate_date == 0)
         {
@@ -1400,8 +1400,8 @@ void map_randsite(int prm_971, int prm_972)
             }
         }
     }
-    if (mdata_map_type >= mdata_t::MapType::dungeon
-        && mdata_map_type <= mdata_t::MapType::dungeon_castle)
+    if (mdata_map_type >= static_cast<int>(mdata_t::MapType::dungeon)
+        && mdata_map_type <= static_cast<int>(mdata_t::MapType::dungeon_castle))
     {
         if (rnd(14) == 0)
         {
